+++ mlx/backend/common/rope.cpp
@@ -1,7 +1,6 @@
 // Copyright © 2023-2024 Apple Inc.
 
-#include "mlx/fast.h"
-#include "mlx/primitives.h"
+#include "mlx/fast_primitives.h"
 
 namespace mlx::core::fast {
